+++ Sources/StoreKitPlus/Services/StoreService.swift
@@ -21,9 +21,10 @@ public protocol StoreService {
     func getProducts() async throws -> [Product]
 
     /// Purchase a certain product.
+    @discardableResult
     func purchase(
         _ product: Product
-    ) async throws -> Product.PurchaseResult
+    ) async throws -> (Product.PurchaseResult, Transaction?)
 
     /// Restore previous purchases.
     @discardableResult
@@ -37,9 +38,38 @@ public protocol StoreService {
 
 public extension StoreService {
 
+    /// Purchase a certain product.
+    ///
+    /// This function will sync the result to the context.
+    @discardableResult
+    func purchase(
+        _ product: Product,
+        syncWith context: StoreContext
+    ) async throws -> (Product.PurchaseResult, Transaction?) {
+        let result = try await purchase(product)
+        if let transaction = result.1 {
+            await context.updatePurchaseTransactions(with: transaction)
+        }
+        return result
+    }
+
+    /// Restore previous purchases.
+    ///
+    /// This function will sync the result to the context.
+    @discardableResult
+    func restorePurchases(
+        syncWith context: StoreContext
+    ) async throws -> [Transaction] {
+        let result = try await restorePurchases()
+        await context.updatePurchaseTransactions(result)
+        return result
+    }
+}
+
+public extension StoreService {
+
     @available(*, deprecated, message: "You have to pass in a context now.")
     func syncStoreData() async throws {
         try await syncStoreData(to: .init())
     }
 }
-